Make reported Unit of Meters configurable
Since the Unit of the meter rates does not have to be time but can be other units as well, this should be configurable and not be limited to the rateUnit of codahale. The rateUnit of codahale does specify the time-part of the unit/time relationship of rate, whereas the Unit of the Cloudwatch metricdatum defines the unit-part.